requirements.txt
#tensorflow==1.4.0
#Keras==2.1.4
matplotlib==2.1.2
numpy==1.14.1
opencv-python==3.2.0.8
Pillow==5.0.0
scikit-image==0.13.1
scipy==1.0.0